# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FormulaWeb QuickBooks Agent",
    "description": "Thin translator between the FormulaWeb ERP and QuickBooks Desktop. JSON in, JSON out.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:8462/"
    }
  ],
  "paths": {
...
    "/customers": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "List customers, optionally only those modified at or after 'updatedSince'. 'maxReturned' caps the result; omit it to fetch all.",
        "operationId": "ListCustomers",
        "parameters": [
          {
            "name": "updatedSince",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "maxReturned",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "Create a customer in QuickBooks. Requires an Idempotency-Key header; replays are safe.",
        "operationId": "CreateCustomer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/customers/{listId}": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Fetch a single customer by its QuickBooks ListID.",
        "operationId": "GetCustomer",
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/items": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "List items (all types), optionally only those modified at or after 'updatedSince'. 'maxReturned' caps the result; omit it to fetch all.",
        "operationId": "ListItems",
        "parameters": [
          {
            "name": "updatedSince",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "maxReturned",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "Create an item in QuickBooks. Requires an Idempotency-Key header; replays are safe.",
        "operationId": "CreateItem",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/items/{listId}": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Fetch a single item by its QuickBooks ListID.",
        "operationId": "GetItem",
        "parameters": [
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateCustomerRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 41,
            "type": "string"
          },
          "companyName": {
            "maxLength": 41,
            "type": [
              "null",
              "string"
            ]
          },
          "firstName": {
            "maxLength": 25,
            "type": [
              "null",
              "string"
            ]
          },
          "lastName": {
            "maxLength": 25,
            "type": [
              "null",
              "string"
            ]
          },
          "phone": {
            "maxLength": 21,
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "maxLength": 1023,
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CreateItemRequest": {
        "required": [
          "type",
          "name"
        ],
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/ItemType"
          },
          "name": {
            "maxLength": 31,
            "type": "string"
          },
          "parentFullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "manufacturerPartNumber": {
            "maxLength": 31,
            "type": [
              "null",
              "string"
            ]
          },
          "salesDescription": {
            "type": [
              "null",
              "string"
            ]
          },
          "purchaseDescription": {
            "type": [
              "null",
              "string"
            ]
          },
          "salesPrice": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "purchaseCost": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "incomeAccountFullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "cogsOrExpenseAccountFullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "assetAccountFullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "reorderPoint": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "openingBalance": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ItemOpeningBalance"
              }
            ]
          }
        }
      },
...
      "CustomerDto": {
        "required": [
          "listId",
          "editSequence",
          "name"
        ],
        "type": "object",
        "properties": {
          "listId": {
            "type": "string"
          },
          "editSequence": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "fullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "companyName": {
            "type": [
              "null",
              "string"
            ]
          },
          "firstName": {
            "type": [
              "null",
              "string"
            ]
          },
          "lastName": {
            "type": [
              "null",
              "string"
            ]
          },
          "phone": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "timeCreated": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "timeModified": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "ItemDto": {
        "required": [
          "listId",
          "editSequence",
          "type",
          "name"
        ],
        "type": "object",
        "properties": {
          "listId": {
            "type": "string"
          },
          "editSequence": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "fullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "parentFullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "manufacturerPartNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "salesDescription": {
            "type": [
              "null",
              "string"
            ]
          },
          "purchaseDescription": {
            "type": [
              "null",
              "string"
            ]
          },
          "salesPrice": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "purchaseCost": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "incomeAccountFullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "cogsOrExpenseAccountFullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "assetAccountFullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "quantityOnHand": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "averageCost": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "reorderPoint": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "timeCreated": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "timeModified": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "ItemOpeningBalance": {
        "required": [
          "quantityOnHand",
          "totalValue",
          "asOf"
        ],
        "type": "object",
        "properties": {
          "quantityOnHand": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalValue": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "asOf": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ItemType": {
        "enum": [
          "Inventory",
          "NonInventory"
        ]
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Shared secret for the internal ERP->agent network.",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
...
}